Moves cli to org app due to project structure [skip ci]
